requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
gradio>=3.0.0
pandas>=1.5.0
requests>=2.28.0
Pillow>=9.0.0
ftfy>=6.1.1
regex>=2023.6.3
tqdm>=4.65.0
git+https://github.com/openai/CLIP.git